Added rune info to navigation and add in empty rune colors for profile
Added the rune info page to the nav, renamed visible wording of
"gem" to "rune" now that things are out and that is what they are
calling it(code still uses gem so couldn't replace to rune everywhere.
Also color coded the empty rune slots on the user_profile page so
that people can get an idea of what type of rune they need for the
missing slots.