Sort by seat now and keep resize red X that is swapped
When moving a crusader down into the form, it would just swap the
images, but the form is 40x40 instead of 48x48. Now it'll resize
the red X to be 48x48. Also sorted by seat, and crusader id, and
wrapped them in a small border so it's somewhat clear who is in
which seat.