New talents added + massive refactor
Added in the newest talents, also re-wrote almost all of the code
to make use the game data a bit more when generating the talents
as well as how the dps is calculated. Still a bit more could be done
to make it cleaner/easier, but I'm tired of looking at/workin on it.