Fix for top to bottom challenge form
They added a campaign order value of 101 for top to bottom as well
as reign of fire. Now I'll be displaying them correctly by checking
for that value instead of just 100. I also had to put a check in
for top to bottom since it defaults to using the DiD form which is
incorrect