Remove workarounds and check for variables before using
They fixed/changed the layout ordering of some talents which made
my workarounds no longer needed. Removed those as well as checked
for a few variables before using them since people are using the
talent picker by manually entering things instead of using the
methods that allow for filling it in for them.