Fix for bonus_training bad suggestions
Since bonus training only does anything once it is at a level above
your max DPS seat, I needed to make sure to set the main dps slot
for it regardless of if you had any levels in it or not. Thanks to
Cry1ng for pointing this bug out.