More player history clean up
Thanks to Zynkster and SepulSoul for asking about this page recently,
I've now added in clean displays for buffs, quests and missions that
give a buff. Still probably a few things missing, but it's getting
closer. Also added in a new function that will give a nice readable
effect given a formatted string of what the effect does.